hacking out a solution
temporary likely, might move much of this up to calvad_rscripts.
However, working on doing the merge here for now.

There is an issue with merging wim and vds, in that I see two extra time
stamps.  Not sure, but I think it is because there are two duplicated
timestamps in the wim.1.N data I am using for testing