fixed processing a site that isn't a "real" site
the site 44 W has data for 2012, but site 44 E does not even though it
is a real site in the DB.  So I was getting a valid distance, even
though there was no imputation.

Fixed by looking at the list of valid wim sites prior to deferring the
action on the wim site.  If the wim site number and direction is not in
the list of valid imputed wim sites for the given year, then it will be
skipped.