didn't like that...trim lanes with more logic
reverted to merging less than everything.  That is, only merge on a new
wim-vds pair if either

1. the wim-vds pair has more lanes of WIM data than the VDS site, or
2. the new wim-vds pair has the same number of lanes as the previously
loaded wim-vds pair

so if you have a VDS site with say 4 lanes, and you try to merge it with
two WIM-VDS pairs sites with just 2 lanes, it will work.  If one WIM-VDS
pair has 3 and the next 2, then just the three lane will get merged.
etc

the idea is to create an imputation that will actually complete!